docs: fixing one example module and generating feature and quickstart docs
