+++ mkdocs.yml
@@ -38,8 +38,16 @@ nav:
   - Home: index.md
   - Quickstart: quickstart.md
   - Features:
-      - features/creating_container.md
-      - features/configuration.md
+      - Creating Containers: features/creating_container.md
+      - Configuration: features/configuration.md
+      - Authentication: features/authentication.md
+      - Executing Commands: features/executing_commands.md
+      - Container Logs: features/container_logs.md
+      - Building Images: features/building_images.md
+      - Copying Data: features/copying_data.md
+      - Wait Strategies: features/wait_strategies.md
+      - Docker Compose: features/docker_compose.md
+      - Networking: features/networking.md
   - Modules:
       - Databases:
           - modules/arangodb.md
