+++ pyproject.toml
@@ -279,6 +279,7 @@ ignore = [
 
 [tool.ruff.lint.per-file-ignores]
 "**/example_*.py" = ["T201"]
+"**/examples/*.py" = ["T201"]
 
 [tool.ruff.lint.pyupgrade]
 keep-runtime-typing = true
